% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_clean}
\alias{sis_clean}
\title{Clean up trait and tree}
\usage{
sis_clean(phy, traits, first_col_names = FALSE)
}
\arguments{
\item{phy}{A phylo object}

\item{traits}{A data.frame of traits}

\item{first_col_names}{Boolean on whether the first column has names.}
}
\value{
a list with phy and traits elements
}
\description{
Does basic formatting and cleanup: makes sure the taxa are the same order in both, makes sure row names of the data are taxa, etc. Relies on geiger's treedata function. The first_col_names is for software like hisse, where the first column is often taxon names.
}
