% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_find_taxon}
\alias{sis_find_taxon}
\title{Is the taxon in one of the sister groups}
\usage{
sis_find_taxon(taxon, sisters)
}
\arguments{
\item{taxon}{Node number of taxon}

\item{sisters}{Data.frame from sis_get_sisters()}
}
\value{
data.frame of whether the taxon is in the left or right sister group, or any
}
\description{
Utility function for tossing out taxa already used
}
