% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_format_comparison}
\alias{sis_format_comparison}
\title{Get comparison format}
\usage{
sis_format_comparison(sisters, trait, phy)
}
\arguments{
\item{sisters}{Data.frame from sis_get_sisters()}

\item{trait}{vector of 0/1 data}

\item{phy}{A phylo object}
}
\value{
data.frame where each row is a sister group comparison.
}
\description{
Convert a data.frame of all sister groups (from sis_get_sisters) and a vector of 0 and 1 (with names equal to taxon names) to a data.frame with the sister groups that differ in traits.
}
\examples{
data(geospiza, package="geiger")
cleaned <- sis_clean(geospiza$phy, geospiza$dat)
phy <- cleaned$phy
traits <- cleaned$traits
trait <- sis_discretize(traits[,1])
sisters <- sis_get_sisters(phy, ncores=2)
sisters_comparison <- sis_format_comparison(sisters, trait, phy)
print(sisters_comparison)
}
