% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_format_simpified}
\alias{sis_format_simpified}
\title{Get simplified comparison format suitable for passing into other functions}
\usage{
sis_format_simpified(sisters_comparison)
}
\arguments{
\item{sisters_comparison}{Data.frame from sis_format_comparison}
}
\value{
Data.frame of two columns: diversity with state 0 and state 1, where each row is a sister group comparison
}
\description{
Get simplified comparison format suitable for passing into other functions
}
