% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_get_monomorphic}
\alias{sis_get_monomorphic}
\title{Get monomorphic trait}
\usage{
sis_get_monomorphic(trait)
}
\arguments{
\item{trait}{Vector of trait values}
}
\value{
The state all taxa have if monomorphic; NA otherwise
}
\description{
Get monomorphic trait
}
