% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_get_sister_pair}
\alias{sis_get_sister_pair}
\title{Get sister groups for a node}
\usage{
sis_get_sister_pair(node, phy)
}
\arguments{
\item{node}{Node number}

\item{phy}{A phylo object}
}
\value{
a data.frame with the node numbers and columns with the tip labels of the two descendant clades
}
\description{
For a node, gives the taxa on each side. Note that the output is a data.frame with lists
}
