% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_get_sisters}
\alias{sis_get_sisters}
\title{Get sister groups for all internal nodes}
\usage{
sis_get_sisters(phy, ncores = 2)
}
\arguments{
\item{phy}{A phylo object}

\item{ncores}{How many cores to use to run this in parallel. I suggest parallel::detectCores(), but set it at 2 for a default (otherwise CRAN checks fail)}
}
\value{
a data.frame with the node numbers and columns with the tip labels of the two descendant clades, plus additional info on the sister groups
}
\description{
For each node, return the vector of tip numbers for taxa on each side. It is sorted so that sister groups with fewer taxa are arranged at the top.
}
