% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{sis_get_trait_values}
\alias{sis_get_trait_values}
\title{Get trait values for tip numbers}
\usage{
sis_get_trait_values(nodes, phy, trait)
}
\arguments{
\item{nodes}{vector of node numbers (tip numbers, actually)}

\item{phy}{A phylo object}

\item{trait}{A trait vector with names equal to taxon names}
}
\description{
Get trait values for tip numbers
}
