% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotest.R
\name{sis_iterate_single_run}
\alias{sis_iterate_single_run}
\title{Do a test with a single cutoff value}
\usage{
sis_iterate_single_run(
  cutoff,
  x,
  use_percentile = TRUE,
  phy,
  sisters = sis_get_sisters(phy),
  drop_matches = TRUE,
  warn = FALSE
)
}
\arguments{
\item{cutoff}{Value to use as cutoff. If percentile, 0.3 = 30th percentile, etc.}

\item{x}{Vector of continuous trait values}

\item{use_percentile}{If TRUE, use cutoff as percentile}

\item{phy}{A phylo object}

\item{sisters}{Data.frame from sis_get_sisters()}

\item{drop_matches}{Drop sister group comparisons with equal numbers of taxa}

\item{warn}{Some tests will fail with warnings (too few sister groups or other reasons). Setting this to FALSE will suppress those}
}
\value{
vector of output from sis_test()
}
\description{
Do a test with a single cutoff value
}
