% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotest.R
\name{sis_test}
\alias{sis_test}
\title{Compute multiple tests based on sister group comparisons}
\usage{
sis_test(pairs, drop_matches = TRUE, warn = TRUE)
}
\arguments{
\item{pairs}{Data.frame with one row per sister group comparison, with one column for number of taxa in state 0, and one column for the number of taxa in state 1.}

\item{drop_matches}{Drop sister group comparisons with equal numbers of taxa}

\item{warn}{Some tests will fail with warnings (too few sister groups or other reasons). Setting this to FALSE will suppress those}
}
\value{
A vector with the results of many tests, as well as summary data for the comparisons
}
\description{
Compute multiple tests based on sister group comparisons
}
\examples{
data(geospiza, package="geiger")
cleaned <- sis_clean(geospiza$phy, geospiza$dat)
phy <- cleaned$phy
traits <- cleaned$traits
trait <- sis_discretize(traits[,1])
sisters <- sis_get_sisters(phy)
sisters_comparison <- sis_format_comparison(sisters, trait, phy)
pairs <- sis_format_simpified(sisters_comparison)
sis_test(pairs)
}
