% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotest.R
\name{tryNA}
\alias{tryNA}
\title{Try but returns NA rather than error}
\usage{
tryNA(code, silent = FALSE)
}
\arguments{
\item{code}{Code to run}

\item{silent}{Print error if TRUE}
}
\description{
Try but returns NA rather than error
}
